+++ src/main/java/insane96mcp/progressivebosses/module/wither/feature/ResistancesFeature.java
@@ -3,6 +3,7 @@ package insane96mcp.progressivebosses.module.wither.feature;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.utils.LogHelper;
 import insane96mcp.progressivebosses.base.Strings;
 import insane96mcp.progressivebosses.setup.Config;
 import net.minecraft.entity.boss.WitherEntity;
@@ -23,7 +24,7 @@ public class ResistancesFeature extends Feature {
 	public double damageReductionBeforeHalfHealth = 0.01d;
 	public double maxDamageReductionBeforeHalfHealth = 0.15d;
 	public double damageReductionOnHalfHealth = 0.02d;
-	public double maxDamageReductionOnHalfHealth = 0.36d;
+	public double maxDamageReductionOnHalfHealth = 0.44d;
 	public double magicDamageBonus = 200d;
 
 	public ResistancesFeature(Module module) {
@@ -77,7 +78,7 @@ public class ResistancesFeature extends Feature {
 
 		CompoundNBT tags = wither.getPersistentData();
 		float difficulty = tags.getFloat(Strings.Tags.DIFFICULTY);
-
+		LogHelper.info("damage before: %s", event.getAmount());
 		//Handle Damage Reduction
 		float damageReduction;
 		if (!wither.isCharged())
@@ -86,5 +87,6 @@ public class ResistancesFeature extends Feature {
 			damageReduction = (float) Math.min(this.maxDamageReductionOnHalfHealth, difficulty * this.damageReductionOnHalfHealth);
 
 		event.setAmount(event.getAmount() * (1f - damageReduction));
+		LogHelper.info("damage after: %s", event.getAmount());
 	}
 }